requirements.txt
Django==1.8.4
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.2.3
Markdown==2.6.2
wheel==0.24.0
